Modify conversation_channel
Change subscribe channel so it uses an unique channel for each user
Unsubscribed method removes all connected connections.
Speak method builds a hash that's based on a passed
data and sends data to the front-end using
ActionCable.serer.broadcast method to the specific channel
Data sent from this method is visible in the received
method on the front-end.