Update message_broadcast_job
The perform method is the main method of this class
The whole code inside of it is executed when you call
the perform_later method. We pass there a message
object, so we can easily get two users to whom a HTML
code should be sent to. Later we'll call the
broadcast_to_sender and the broadcast_to_recipient
methods. For now, they look the same but we will extend
the broadcast_to_recipient later